requirements.txt
base58
solana
python-telegram-bot
supabase